+++ libs/portal-integration-angular/src/lib/core/components/interactive-data-view/interactive-data-view.component.ts
@@ -4,7 +4,10 @@ import { DataSortDirection } from '../../../model/data-sort-direction'
 import { Filter, Row, Sort } from '../data-table/data-table.component'
 import { DataViewComponent, RowListGridData } from '../data-view/data-view.component'
 import { GroupSelectionChangedEvent } from '../column-group-selection/column-group-selection.component'
-import { ActionColumnChangedEvent, ColumnSelectionChangedEvent } from '../custom-group-column-selector/custom-group-column-selector.component'
+import {
+  ActionColumnChangedEvent,
+  ColumnSelectionChangedEvent,
+} from '../custom-group-column-selector/custom-group-column-selector.component'
 import { DataAction } from '../../../model/data-action'
 
 @Component({
@@ -54,7 +57,9 @@ export class InteractiveDataViewComponent implements OnInit {
   @Input() additionalActions: DataAction[] = []
   @Input() listGridPaginator = true
   @Input() tablePaginator = true
+  @Input() page = 0
   @Input() selectedRows: Row[] = []
+  @Input() displayedColumns: DataTableColumn[] = []
   @ContentChild('tableCell') tableCell: TemplateRef<any> | undefined
   @ContentChild('tableDateCell') tableDateCell: TemplateRef<any> | undefined
   @ContentChild('tableRelativeDateCell') tableRelativeDateCell: TemplateRef<any> | undefined
@@ -74,7 +79,8 @@ export class InteractiveDataViewComponent implements OnInit {
   @Output() dataViewLayoutChange = new EventEmitter<'grid' | 'list' | 'table'>()
   @Output() displayedColumnsChange = new EventEmitter<DataTableColumn[]>()
   @Output() selectionChanged: EventEmitter<Row[]> = new EventEmitter()
-  displayedColumns: DataTableColumn[] = []
+
+  @Output() pageChanged: EventEmitter<number> = new EventEmitter()
   selectedGroupKey = ''
   isDeleteItemObserved: boolean | undefined
   isViewItemObserved: boolean | undefined
@@ -222,8 +228,8 @@ export class InteractiveDataViewComponent implements OnInit {
         })
       }
     }
-    if(this.selectionChanged.observed) {
-      if(!this._dataViewComponent?.selectionChanged.observed) {
+    if (this.selectionChanged.observed) {
+      if (!this._dataViewComponent?.selectionChanged.observed) {
         this._dataViewComponent?.selectionChanged.subscribe((event) => {
           this.onRowSelectionChange(event)
         })
@@ -243,8 +249,13 @@ export class InteractiveDataViewComponent implements OnInit {
   }
 
   onRowSelectionChange(event: Row[]) {
-    if(this.selectionChanged.observed){
+    if (this.selectionChanged.observed) {
       this.selectionChanged.emit(event)
     }
+  }
+
+  onPageChange(event: number) {
+    this.page = event
+    this.pageChanged.emit(event)
   }
 }