+++ libs/portal-integration-angular/src/lib/core/components/lifecycle/lifecycle.component.ts
@@ -1,0 +1,16 @@
+import { Component, Input } from '@angular/core'
+
+export interface LifecycleStep {
+  id: string
+  title: string,
+  details?: string,
+}
+
+@Component({
+  selector: 'ocx-lifecycle',
+  templateUrl: './lifecycle.component.html',
+})
+export class LifecycleComponent {
+  @Input() steps: LifecycleStep[] = []
+  @Input() activeStepId: string | undefined
+}