+++ libs/portal-integration-angular/testing/diagram.harness.ts
@@ -1,5 +1,6 @@
 import { ComponentHarness } from '@angular/cdk/testing'
 import { PChartHarness } from './primeng/p-chart.harness'
+import { PSelectButtonHarness } from './primeng/p-selectButton.harness'
 
 export class DiagramHarness extends ComponentHarness {
   static hostSelector = 'ocx-diagram'
@@ -12,5 +13,13 @@ export class DiagramHarness extends ComponentHarness {
 
   async getSumLabel(): Promise<string | undefined> {
     return (await this.locatorForOptional('.sumKey span[name="sumLabel"]')())?.text()
+  }
+
+  async getDiagramTypeSelectButton() {
+    return (await this.locatorForOptional('p-selectButton[name="diagram-type-select-button"]')())
+  }
+
+  async getAllSelectionButtons() {
+    return await (await this.locatorFor(PSelectButtonHarness)()).getAllButtons()
   }
 }